ocr/requirements.txt
numpy==1.16.1
opencv-contrib-python==4.0.0.21
Pillow==5.4.1
pkg-resources==0.0.0
pytesseract==0.2.6
